Restructured the code for new contact pane and form
